Unable to link the CSS stylesheet. Therefore put the styling in the html page. Also edited the detail goal page, to try if it does not work there either. No clue how to fix it
Created an issue for it on github:
"Unable to link the css stylesheet in the html pages.
I ran the collectstatic. However, MissingValueError is still raised. The stylesheet cannot be found, although FindStatic is able to find the stylesheet. The CSS file from sociable_detail does work.
Without the static tag it does not raise a server error (500) anymore. However, it is still unable to find the css stylesheet.
Putting the link in the main template also raises a server error (500)

Putting the styling within the HTML does work.

Anyone can help?"