Fixed the but that the stylesheet cannot be found. This is due to a whitenoise compression. See https://stackoverflow.com/questions/26829435/collectstatic-command-fails-when-whitenoise-is-enabled/32347324#32347324
What I can do better next time to find the cause of the bug: Find everything that could be related to the bug. I knew it had something to do with static files, but I did not know what. If I searched all files with the name static, then I would have found that there is a setting with serving static files. This was the problem