Big commit. Followed the tutorials https://simpleisbetterthancomplex.com/tutorial/2016/07/22/how-to-extend-django-user-model.html#onetoone to extend user model with a profile model I tried to use a class based view to make the edit account form. However, I did not succeed, as I got stuck with the unit testing of the forms. I used the update view of the tutorial
Also tried to change the auto_login fixture, to be able to shorten the code. Did not succeed in this either. Did remove the create_user fixture from the user tests and used the UserFactory for that.

WARNING: Due to this commit, the database is broken. A new database needs to be initialized