#102 Chatmessages are now encrypted when saving and decrypted when in chatroom. For a few seconds you see the encrypted chatmessage when you save it. It is stored encrypted in the database.
I had to delete the update functionality for chat messages. Anonymous users can delete and rewrite the chatmessage. Got too much trouble with updating chat messages.

As a logged in user you are not able to send messages. Only as anonymous user. As a logged in user, the message is not saved. Do not know what is wrong there